refactor the main.rb
